# Problem: contract call which does lots of message calls is slow to execute