<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰    
v    Before smashing the submit button please review the checkboxes.
v    If a checkbox is n/a - please still include it but + a little note why
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

> This PR is huge, added some [guide for review](#issuecomment-971114131) in the comment section below.

## Description

Closes: #710 

As profiled in the issue page, the bottleneck is the slowness of deep context stack, there's no easy way around it other than refactor the `StateDB` to use journal logs like what go-ethereum does, thus this big code refactoring.

## What this PR does

- Implement a standalone statedb module in `x/evm/statedb` which implements `vm.StateDB` interface with the support of several keeper methods, the keeper interface is captured in `statedb.Keeper`.
- The `StateDB` implementation keep all the dirty states in memory, at the end of tx execution, commit all the dirty changes to keeper at once
- record all the modify operations in a list of journal logs, snapshot revert is done by undo part of the logs.
- `x/evm/statedb/{journal.go, access_list.go, state_object.go}` are directly adapted from go-ethereum itself.

## Benchmark result

### BenchmarkMessageCall (improved 310X)

#### Before

```
BenchmarkMessageCall-16    	       1	25846269634 ns/op
BenchmarkMessageCall-16    	       1	26773165908 ns/op
BenchmarkMessageCall-16    	       1	25786298762 ns/op
```

#### After

```
BenchmarkMessageCall-16    	      12	  91367037 ns/op
BenchmarkMessageCall-16    	      13	  82284757 ns/op
BenchmarkMessageCall-16    	      14	  83061467 ns/op
```

### BenchmarkEmitLogs (improved 2.5X)

#### Before

```
BenchmarkEmitLogs-16    	      98	  12034965 ns/op
BenchmarkEmitLogs-16    	      98	  12146658 ns/op
BenchmarkEmitLogs-16    	      96	  12313053 ns/op
```

#### After

```
BenchmarkEmitLogs-16    	     225	   5288751 ns/op
BenchmarkEmitLogs-16    	     232	   5189175 ns/op
BenchmarkEmitLogs-16    	     222	   5191943 ns/op
```


<!-- Add a description of the changes that this PR introduces and the files that
are the most critical to review.
-->

______

For contributor use:

- [ ] Targeted PR against correct branch (see [CONTRIBUTING.md](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#pr-targeting))
- [ ] Linked to Github issue with discussion and accepted design OR link to spec that describes this work.
- [ ] Code follows the [module structure standards](https://github.com/cosmos/cosmos-sdk/blob/master/docs/building-modules/structure.md).
- [ ] Wrote unit and integration [tests](https://github.com/cosmos/cosmos-sdk/blob/master/CONTRIBUTING.md#testing)
- [ ] Updated relevant documentation (`docs/`) or specification (`x/<module>/spec/`)
- [ ] Added  relevant `godoc` [comments](https://blog.golang.org/godoc-documenting-go-code).
- [ ] Added a relevant changelog entry to the `Unreleased` section in `CHANGELOG.md`
- [ ] Re-reviewed `Files changed` in the Github PR explorer

______

For admin use:

- [ ] Added appropriate labels to PR (ex. `WIP`, `R4R`, `docs`, etc)
- [ ] Reviewers assigned
- [ ] Squashed all commits, uses message "Merge pull request #XYZ: [title]" ([coding standards](https://github.com/tendermint/coding/blob/master/README.md#merging-a-pr))