@yihuang 2021-11-08
> @yihuang could I kindly ask if we've considered that things might be much faster using smt inside of cosmos-sdk instead of cosmos/iavl?

Does that mean we need to implement state-sync again?