@tomtau 2021-11-09
> @yihuang could I kindly ask if we've considered that things might be much faster using smt inside of cosmos-sdk instead of cosmos/iavl?

@odeke-em SMT is orthogonal to it AFAIK: one of the core issues here is e.g. nested cache contexts (https://github.com/cosmos/cosmos-sdk/issues/10310) https://github.com/tharsis/ethermint/issues/710#issuecomment-957101207