@tomtau 2021-11-16
the importer test runs faster now:
```
processed block: 97000 (time so far: 3m3.791587401s)
--- PASS: TestImporterTestSuite (186.34s)
    --- PASS: TestImporterTestSuite/TestImportBlocks (186.34s)
```

vs on main:
```
processed block: 97000 (time so far: 6m53.091837321s)
--- PASS: TestImporterTestSuite (416.45s)
    --- PASS: TestImporterTestSuite/TestImportBlocks (416.45s)
```