@yihuang 2021-11-17
# Review Guide

## Abstract

The problem:

- Deep context stack is slow to do commit, and it's hard to workaround.
- Have to call `WithContext` on `EvmKeeper` is error prone.

Basic idea is adopt the design in go-ethereum:

- Cache the intermediate states during contract execution in memory, flush the dirty states to storage at the end of tx execution.

- Move the `StateDB` implementation out of `EvmKeeper`, then the `EvmKeeper` mostly only provides stateless operations on chain storage.
- snapshot and revert is done by record and undo the journal logs.

## `x/evm/statedb` package

All the statedb logic is implemented in this package, the access to underlying storage is provided by `statedb.Keeper` interface, which is implemented by `EvmKeeper`.

Most of the `StateDB` implementation are directly adapted from go-ethereum codebase itself.

## Transient Storage Simplification

The transient storage is greatly simplified since many of temparary states are moved to memory, the left ones are:

- `KeyPrefixTransientBloom` store current block bloom
- `KeyPrefixTransientTxIndex` store current tx index
- `KeyPrefixTransientLogSize` store current log index

## Persistent Storage

It don't change the format of persistent storage, so no migration needed when upgrade. But because of different storage writing order, the app hash would be different.