@leejw51crypto 2021-11-30
could you add some comment why dramatic speed enhancement is achieved?