@yihuang 2021-11-30
> could you add some comment why dramatic speed enhancement is achieved? also whether it's consensus breaking change, how to resolve it in upgrading

https://github.com/tharsis/ethermint/pull/729#issuecomment-971114131
it's explained in the review guide, it's consensus breaking, but don't need storage migration, only root hashes changed.