@yihuang 2021-11-30
About potential overflow/underflow in `AddBalance` and `SubBalance`, I think it's not possible in reality because of EVM logic.
I grepped through the go-ethereum source code, the only two relevant places are `Transfer` and `opSuicide`, both maintain the invariant of the fixed total supply.

```shell
go-ethereum/core/ v1.10.11 $ rg -w "AddBalance|SubBalance"
evm.go
117:	db.SubBalance(sender, amount)
118:	db.AddBalance(recipient, amount)

* This is a `Transfer` function, should be the main function to be called to handle balance change, it maintains invariant of fixed total supply.

state_transition.go
210:	st.state.SubBalance(st.msg.From(), mgval)
336:	st.state.AddBalance(st.evm.Context.Coinbase, new(big.Int).Mul(new(big.Int).SetUint64(st.gasUsed()), effectiveTip))
355:	st.state.AddBalance(st.msg.From(), remaining)

# They are all gas fee logic, we didn't reuse this part of code.

genesis.go
267:		statedb.AddBalance(addr, account.Balance)

# Genesis logic, not relevant

state/state_object.go
412:// AddBalance adds amount to s's balance.
414:func (s *stateObject) AddBalance(amount *big.Int) {
426:// SubBalance removes amount from s's balance.
428:func (s *stateObject) SubBalance(amount *big.Int) {

# Implementation, not relevant

vm/evm.go
359:	// We do an AddBalance of zero here, just in order to trigger a touch.
363:	evm.StateDB.AddBalance(addr, big0)

# A special logic to add zero balance, not relevant

vm/interface.go
30:	SubBalance(common.Address, *big.Int)
31:	AddBalance(common.Address, *big.Int)

# Interface definition, not relevant

vm/instructions.go
792:	interpreter.evm.StateDB.AddBalance(beneficiary.Bytes20(), balance)

# opSuicide, add the balance of contract address to beneficiary, the balance of contract address is burn in `Suicide` method. Still maintain the fixed total supply.

state/statedb.go
377:// AddBalance adds amount to the account associated with addr.
378:func (s *StateDB) AddBalance(addr common.Address, amount *big.Int) {
381:		stateObject.AddBalance(amount)
385:// SubBalance subtracts amount from the account associated with addr.
386:func (s *StateDB) SubBalance(addr common.Address, amount *big.Int) {
389:		stateObject.SubBalance(amount)

# Implementation, not relevant
```