@crypto-facs 2021-12-01
LGTM. I have ran all the tests and did some manual testing my self and RPCs are working as expected. Ran against blockscout as well to make sure I was not missing anything, but there were no issues. From the RPC side this seems like a safe change.