@yihuang 2022-01-05
> Final rounds of comments. Mostly adding more code and godoc comments. 

for the unit tests and coding style of statedb package, I suggest postponing that to another PR (https://github.com/tharsis/ethermint/issues/876), because:
- this PR is big enough
- there's are several modules under `x/evm/statedb` that are copied from go-ethereum, the diffs are small, it's easy to track what have we changed on top of go-ethereum version.

> Can you also add a state machine breaking change changelog? 

done

> Also can you open a PR to update the spec with the new design?

https://github.com/tharsis/ethermint/issues/879